Migration of main routines to mdmain.cpp to increase modularity. In the future tinyxml2 might be ditched and I might start using libspud.
