+++ examples/GCM_phase_diagram/gcm_compression.cpp
@@ -5,7 +5,7 @@
 #include "phase_transition.h"
 
 int main() {
-  size_t compress_steps = 5000;
+  size_t compress_steps = 1000;
 
   /*
    * This method effectively replaces timestep iterations with
@@ -13,16 +13,13 @@ int main() {
    *
    */
   phase_transition run(compress_steps, 7, "FCC");
-  run.set_compression_flag(true);
 
-  std::vector<double> temperatures = {
-      4.602932914279152937e-05, 6.461042397802770065e-05,
-      6.454528429411781482e-05, 7.688925439500345071e-05,
-      8.923322449588821925e-05, 1.078306042521020788e-04};
+  double rho_start = 1.0e-01;
+  double rho_end = 2.0e-01;
+  std::vector<double> t2 = helper_functions::linspace(5.0e-04, 3.0e-03, 20);
 
-  for (size_t t = 0; t < temperatures.size(); ++t) {
-    run.crystallisation("compress_fwd_" + std::to_string(t), 3.40e-02, 4.2e-02,
-                        double(abs(3.40e-02 - 4.2e-02) / 8),
-                        temperatures[t], NAN, NAN, "GCM");
+  for (size_t t = 0; t < t2.size(); ++t) {
+    run.crystallisation("compress_fwd_" + std::to_string(t), rho_start, rho_end,
+                        (rho_end - rho_start) / 20, t2[t], NAN, NAN, "GCM");
   }
 }
